app/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
